+++ apps/api/config/app.php
@@ -27,7 +27,7 @@ return [
     |
     */
 
-    'version' => '0.22.0', // x-release-please-version
+    'version' => '0.23.0', // x-release-please-version
 
     /*
     |--------------------------------------------------------------------------